Refactored: move cookie helper from composables to helpers
